add customization for edit section control
