+++ packages/dashboard/src/components/Section/SectionPage.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react';
 import { ExampleCard, IComponentDemoPageProps, ComponentPage, PropertiesTableSet } from '@uifabric/example-app-base';
 import { EditSectionsExample } from './examples/EditSections.Example';
+import { EditSectionsCustomizationExample } from './examples/EditSections.Customization.Example';
 const EditSectionsExampleCode = require('!raw-loader!@uifabric/dashboard/src/components/Section/examples/EditSections.Example.tsx') as string;
+const EditSectionsCustomizationExampleCpde = require('!raw-loader!@uifabric/dashboard/src/components/Section/examples/EditSections.Customization.Example.tsx') as string;
 
 export class SectionPage extends React.Component<IComponentDemoPageProps, {}> {
   public render(): JSX.Element {
@@ -14,12 +16,13 @@ export class SectionPage extends React.Component<IComponentDemoPageProps, {}> {
             <ExampleCard title="Edit Sections" isScrollable={true} isOptIn={true} code={EditSectionsExampleCode}>
               <EditSectionsExample />
             </ExampleCard>
+            <ExampleCard title="Edit Sections Customization" isScrollable={true} isOptIn={true} code={EditSectionsCustomizationExampleCpde}>
+              <EditSectionsCustomizationExample />
+            </ExampleCard>
           </div>
         }
         propertiesTables={
-          <PropertiesTableSet
-            sources={[require<string>('!raw-loader!@uifabric/dashboard/src/components/Section/Section.types.ts')]}
-          />
+          <PropertiesTableSet sources={[require<string>('!raw-loader!@uifabric/dashboard/src/components/Section/Section.types.ts')]} />
         }
         overview={<div>Sections and edit sections</div>}
         bestPractices={<div />}
