+++ packages/dashboard/src/components/Section/index.ts
@@ -1,3 +1,4 @@
 export * from './Section';
 export * from './Section.types';
 export * from './EditSections';
+export * from './EditSections.base';
